Removes unnecesary telemetry dependency.
